Create a seperate class PersonCreator
