+++ main.rb
@@ -1,18 +1,14 @@
 # rubocop:disable Metrics/CyclomaticComplexity
 require './book'
-require './person'
-require './student'
-require './teacher'
 require './classroom'
-require 'colorize'
+require './person_creator'
 
 class Main
-  attr_accessor :book_list, :people_list, :rental_list
+  attr_accessor :book_list, :rental_list
 
   def initialize
     @book_list = []
-    @people_list = []
-    @rental_list = []
+    @person_creator = PersonCreator.new
   end
 
   def start
@@ -27,8 +23,8 @@ class Main
   def execute_option(option)
     case option
     when '1' then show_book_list
-    when '2' then show_people_list
-    when '3' then create_person
+    when '2' then @person_creator.show_people_list
+    when '3' then @person_creator.create_person
     when '4' then create_book
     when '5' then create_rental
     when '6' then show_rentals_for_person
@@ -37,28 +33,6 @@ class Main
   end
 
   # rubocop:enable Metrics/CyclomaticComplexity
-
-  def create_person
-    print 'Do you want to create a student (1) or a teacher (2)? [Input the number]: '.blue
-    status = gets.chomp
-    print 'Age: '.blue
-    age = gets.chomp
-    print 'Name: '.blue
-    name = gets.chomp
-    if status == '1'
-      print 'Has a parent permission? [Y/N] '.blue
-      permission = gets.chomp.downcase
-      @people_list << Student.new(age, name, parent_permission: permission == 'y')
-      puts 'Student has been created successfully'.green
-    else
-      print 'Specialization: '.blue
-      spec = gets.chomp
-      @people_list << Teacher.new(spec, age, name)
-      puts 'Teacher has been created successfully'.green
-    end
-
-    start
-  end
 
   def create_book
     print 'Title: '.blue
@@ -83,18 +57,6 @@ class Main
     start
   end
 
-  def show_people_list
-    if @people_list.empty?
-      puts '( No People Found )'.red
-    else
-      @people_list.each_with_index.map do |person, index|
-        puts "#{index})[#{person.class.name}] Name: #{person.name}, ID: #{person.id}, Age: #{person.age}".yellow
-      end
-    end
-
-    start
-  end
-
   def create_rental
     puts 'Select a book from the following list by number'.blue
     @book_list.each_with_index.map do |book, index|
@@ -102,7 +64,7 @@ class Main
     end
     index_book = gets.chomp.to_i
     puts 'Select a person from the following list by number'.blue
-    @people_list.each_with_index.map do |person, index|
+    @person_creator.people_list.each_with_index.map do |person, index|
       puts "#{index})[#{person.class.name}] Name: #{person.name}, ID: #{person.id}, Age: #{person.age}".yellow
     end
     index_person = gets.chomp.to_i
@@ -110,7 +72,7 @@ class Main
     rent_date = gets.chomp
     rental = Rental.new(rent_date)
     @book_list[index_book].add_rental(rental)
-    @people_list[index_person].add_rental(rental)
+    @person_creator.people_list[index_person].add_rental(rental)
     @rental_list << rental
     puts 'Rental has been created successfully'.green
     start
@@ -119,7 +81,7 @@ class Main
   def show_rentals_for_person
     print 'ID of person: '.blue
     person_id = gets.chomp.to_i
-    person = @people_list.select { |target| target.id == person_id }[0]
+    person = @person_creator.people_list.select { |target| target.id == person_id }[0]
     person.rentals.map { |rent| puts "Data: #{rent.date}, Book: \"#{rent.book.title}\", by #{rent.book.author}".yellow }
     start
   end
